+++ staypresent/__init__.py
@@ -19,5 +19,5 @@ from .runner import run
 from .pinger import ping, cron, active_cron_handles
 from . import web
 
-__version__ = "1.5.14"
+__version__ = "1.5.16"
 __all__ = ["run", "web", "ping", "cron", "active_cron_handles", "__version__"]